+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Next release
 * Add Varanus::SSL#report
+* Varanus::Reports (Varanus#reports) is now deprecated.
 
 ### Version 0.5.1 (2021-01-28)
 * Varanus::SSL::CSR - support EC certs