+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### Next release
+### Version 0.6.0 (2020-02-01)
 * Add Varanus::SSL#report
 * Varanus::Reports (Varanus#reports) is now deprecated.